+++ lib/src/tiny_mce_editor.dart
@@ -1,0 +1,2 @@
+export 'tiny_mce_editor_stub.dart'
+    if (dart.library.html) 'tiny_mce_editor_web.dart';