+++ lib/src/tiny_mce_editor_widget.dart
@@ -1,2 +1,0 @@
-export 'tiny_mce_editor_widget_stub.dart'
-    if (dart.library.html) 'tiny_mce_editor_widget_web.dart';