+++ lib/tiny_flutter.dart
@@ -1,1 +1,1 @@
-export 'src/tiny_mce_editor_widget.dart';
+export 'src/tiny_mce_editor.dart';